+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - Inherit font-family
 - Change containers in `inline-flex`
 - Added double-border class (db)
-- Set color from prop - Button
+- Set `color` & `fontColor` from prop - Button
 - Set default border radius to `2px`
 - Added `caret` in Select
 - Fixed Select dropdown padding